Move weight factors calculations out of the controller to their own class
