+++ sources/controllers/Search.controller.php
@@ -24,24 +24,6 @@
 class Search_Controller extends Action_Controller
 {
 	/**
-	 * Weighing factor each area, ie frequency, age, sticky, etc
-	 * @var array
-	 */
-	protected $_weight = array();
-
-	/**
-	 * Holds the total of all weight factors, should be 100
-	 * @var int
-	 */
-	protected $_weight_total = 0;
-
-	/**
-	 * Holds array of search result relevancy weigh factors
-	 * @var array
-	 */
-	protected $_weight_factors = array();
-
-	/**
 	 * Holds the search object
 	 * @var \ElkArte\Search\Search
 	 */
@@ -263,9 +245,6 @@ class Search_Controller extends Action_Controller
 		// No, no, no... this is a bit hard on the server, so don't you go prefetching it!
 		stop_prefetching();
 
-		// Set up the weights to help tune result relevancy
-		$this->_setup_weight_factors();
-
 		// These vars don't require an interface, they're just here for tweaking.
 		$recentPercentage = 0.30;
 		$humungousTopicPosts = 200;
@@ -292,7 +271,7 @@ class Search_Controller extends Action_Controller
 		isAllowedTo('search_posts');
 
 		$this->_search = new \ElkArte\Search\Search($humungousTopicPosts, $maxMessageResults);
-		$this->_search->setWeights($this->_weight_factors, $this->_weight, $this->_weight_total);
+		$this->_search->setWeights(new \ElkArte\Search\WeightFactors($modSettings, $user_info['is_admin']));
 
 		if (isset($_REQUEST['params']))
 			$this->_search->searchParamsFromString($_REQUEST['params']);
@@ -766,94 +745,6 @@ class Search_Controller extends Action_Controller
 	private function _highlighted_callback($matches)
 	{
 		return isset($matches[2]) && $matches[2] == $matches[1] ? stripslashes($matches[1]) : '<span class="highlight">' . $matches[1] . '</span>';
-	}
-
-	/**
-	 * Prepares the weighting factors
-	 */
-	private function _setup_weight_factors()
-	{
-		global $user_info, $modSettings;
-
-		$default_factors = $this->_weight_factors = array(
-			'frequency' => array(
-				'search' => 'COUNT(*) / (MAX(t.num_replies) + 1)',
-				'results' => '(t.num_replies + 1)',
-			),
-			'age' => array(
-				'search' => 'CASE WHEN MAX(m.id_msg) < {int:min_msg} THEN 0 ELSE (MAX(m.id_msg) - {int:min_msg}) / {int:recent_message} END',
-				'results' => 'CASE WHEN t.id_first_msg < {int:min_msg} THEN 0 ELSE (t.id_first_msg - {int:min_msg}) / {int:recent_message} END',
-			),
-			'length' => array(
-				'search' => 'CASE WHEN MAX(t.num_replies) < {int:huge_topic_posts} THEN MAX(t.num_replies) / {int:huge_topic_posts} ELSE 1 END',
-				'results' => 'CASE WHEN t.num_replies < {int:huge_topic_posts} THEN t.num_replies / {int:huge_topic_posts} ELSE 1 END',
-			),
-			'subject' => array(
-				'search' => 0,
-				'results' => 0,
-			),
-			'first_message' => array(
-				'search' => 'CASE WHEN MIN(m.id_msg) = MAX(t.id_first_msg) THEN 1 ELSE 0 END',
-			),
-			'sticky' => array(
-				'search' => 'MAX(t.is_sticky)',
-				'results' => 't.is_sticky',
-			),
-			'likes' => array(
-				'search' => 'MAX(t.num_likes)',
-				'results' => 't.num_likes',
-			),
-		);
-
-		// These are fallback weights in case of errors somewhere.
-		// Not intended to be passed to the hook
-		$default_weights = array(
-			'search_weight_frequency' => 30,
-			'search_weight_age' => 25,
-			'search_weight_length' => 20,
-			'search_weight_subject' => 15,
-			'search_weight_first_message' => 10,
-		);
-
-		call_integration_hook('integrate_search_weights', array(&$this->_weight_factors));
-
-		// Set the weight factors for each area (frequency, age, etc) as defined in the ACP
-		$this->_calculate_weights($this->_weight_factors, $modSettings);
-
-		// Zero weight.  Weightless :P.
-		if (empty($this->_weight_total))
-		{
-			// Admins can be bothered with a failure
-			if ($user_info['is_admin'])
-				throw new Elk_Exception('search_invalid_weights');
-
-			// Even if users will get an answer, the admin should know something is broken
-			Errors::instance()->log_lang_error('search_invalid_weights');
-
-			// Instead is better to give normal users and guests some kind of result
-			// using our defaults.
-			// Using a different variable here because it may be the hook is screwing
-			// things up
-			$this->_calculate_weights($default_factors, $default_weights);
-		}
-	}
-
-	/**
-	 * Fill the $_weight variable and calculate the total weight
-	 *
-	 * @param mixed[] $factors
-	 * @param int[] $weights
-	 */
-	private function _calculate_weights($factors, $weights)
-	{
-		$this->_weight = array();
-		$this->_weight_total = 0;
-
-		foreach ($factors as $weight_factor => $value)
-		{
-			$this->_weight[$weight_factor] = empty($weights['search_weight_' . $weight_factor]) ? 0 : (int) $weights['search_weight_' . $weight_factor];
-			$this->_weight_total += $this->_weight[$weight_factor];
-		}
 	}
 
 	/**
